pyproject.toml
[project]
name = "assignment-crewai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai==0.80.0",
    "langchain-groq==0.2.4",
    "pdfplumber==0.11.4",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "setuptools==69.5.1",
    "streamlit==1.41.1",
]
